% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{aggregate_hourly}
\alias{aggregate_hourly}
\title{aggregate_hourly}
\usage{
aggregate_hourly(fires)
}
\arguments{
\item{fires}{A data frame of fires as returned from \code{\link{extract_fires}}}
}
\value{
A data frame of hourly aggregate fires
}
\description{
A simple convenience function to turn raw extracted fires into hourly aggregates. Area
is summed, power and temperature are averaged, and all mask values are listed. If FRE
and TPM emissions are present, they are summed as well.
}
\examples{
hourly_fires <- aggregate_hourly(fires)
}
