% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofiev.R
\name{assign_met_files}
\alias{assign_met_files}
\title{Assign each hour in the fire list with a specific met file and hour}
\usage{
assign_met_files(fires, met_path)
}
\arguments{
\item{fires}{A data frame of hourly fire information to process with Hour in UTC}

\item{met_path}{A path to WRF netcdf files with the parameters needed for Sofiev (PBLH,
PH, PHB, T, HGT, XLAT, XLONG)}
}
\value{
Returns the same data frame of fires with a filename of met data attached
}
\description{
Assign each hour in the fire list with a specific met file and hour
}
