% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_process}
\alias{batch_process}
\title{batch_process}
\usage{
batch_process(url, download_path, output_path, outname)
}
\arguments{
\item{url}{Subscription FTP location (e.g.,
"ftp://ftp.class.ncdc.noaa.gov/sub/sraffuse1/52654/")}

\item{download_path}{Local path where netcdf files are stored}

\item{output_path}{Local path where output csv will be written}

\item{outname}{Name of output csv. The script will add a time stamp}
}
\value{
A data frame that matches the output csv
}
\description{
Used for routine acquisition of GOES FDCC files from a CLASS subscription. The function
(really a script) will open an ftp location, find all of the netcdf FDCC files
(currently skipping FDCF), and process all of those files that are not already found in
the download_path. The new files will be downloaded and processed, with the results
concatenated into a sinlge csv output file
}
\examples{
df <- batch_process("ftp://ftp.class.ncdc.noaa.gov/sub/sraffuse1/52654/",
 "../../data/netcdf/", "../../data/csv/", "GOES_Fires")
}
