% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofiev.R
\name{calculate_plume_top}
\alias{calculate_plume_top}
\title{Calculate hourly plume top heights using the method of Sofiev et al (2012) and met data
from WRF}
\usage{
calculate_plume_top(fires, met_path)
}
\arguments{
\item{fires}{data frame of fire locations produced by get_pthourly_frp}

\item{met_path}{path for exported met files}
}
\description{
Calculate hourly plume top heights using the method of Sofiev et al (2012) and met data
from WRF
}
