% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube_subset}
\alias{cube_subset}
\title{Refine a goesfire cube to only the points that fall within a polygon}
\usage{
cube_subset(cube, polygon, buffer = 4, crs = 5070)
}
\arguments{
\item{cube}{A goesfire cube data.frame as produced by \code{\link{create_cube}}}

\item{polygon}{A simple feature polygon (or multipolygon)}

\item{buffer}{Buffer distance in km (default = 4)}

\item{crs}{Coordinate reference system: integer with EPSG code, or character with
proj4string (default is 5070 - Albers Equal Area)}
}
\value{
A goesfire cube data.frame, subsetted by the polygon
}
\description{
Refine a goesfire cube to only the points that fall within a polygon
}
