% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{extract_fires}
\alias{extract_fires}
\title{extract_fires}
\usage{
extract_fires(filename, vars = c("Temp", "Power", "Area", "DQF", "Mask"),
  maskvals = NULL)
}
\arguments{
\item{filename}{the full path name of the netcdf FDC file to process}

\item{vars}{The variables to include in the output among c("Temp", "Power", "Area",
"DQF", "Mask"). All are included by default.}

\item{maskvals}{If "Mask" is selected in vars, these are the mask values to keep. If no
mask values are selected, all pixels will be returned and the output will be very
large.}
}
\value{
A data frame with latitude, longitude, time, and the variables requested.
}
\description{
Pull fire information from GOES-R ABI netcdf files and return a csv with lat/lon info
}
\examples{
df <- extract_fires(infile, maskvals = c(10, 11, 30, 31))
}
