% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{feer_emissions}
\alias{feer_emissions}
\title{feer_emissions}
\usage{
feer_emissions(fires, feer)
}
\arguments{
\item{fires}{A data frame of fires processed by \code{extract_fires}}

\item{feer}{The FEERv1.0_Ce.csv file loaded as a data frame. This can be acquired from
\url{https://feer.gsfc.nasa.gov/}}
}
\value{
A data frame with FEER information and TPM (in kg) attached
}
\description{
Estimate emissions of total PM using the smoke emissions coefficient lookup developed
by the NASA Fire Energetics and Emissions Research group. If a Ce value is missing, we
use a fill value of 0.0159 kg/MJ, which is the global median. Also calculates total
heat assuming a radiant fraction and PM2.5 assuming a ratio of PM2.5 to TPM
}
\examples{
emissions <- feer_emissions(fires, feer)
}
