% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_bluesky_fire_locations}
\alias{get_bluesky_fire_locations}
\title{Take hourly activity, such as from \code{\link{get_pthourly_frp}} or
\code{\link{get_hourly_profile}} and return BlueSky fire_locations at daily resolution}
\usage{
get_bluesky_fire_locations(df, fire_name, final_area)
}
\arguments{
\item{final_area}{}
}
\description{
Take hourly activity, such as from \code{\link{get_pthourly_frp}} or
\code{\link{get_hourly_profile}} and return BlueSky fire_locations at daily resolution
}
