% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coord_bounds}
\alias{get_coord_bounds}
\title{get_coord_bounds}
\usage{
get_coord_bounds(nc)
}
\arguments{
\item{nc}{}
}
\value{
A named list with x1, x2, y1, and y2
}
\description{
Determines the x and y bounds (in radians) for a given netcdf handle
}
