% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_SMOKE}
\alias{output_SMOKE}
\title{Create SMOKE-ready files (ems95 and ida) for use in SMOKE-CMAQ}
\usage{
output_SMOKE(fires, path)
}
\arguments{
\item{fires}{data frame of hourly fire emissions as produced by
\code{\link{feer_emissions}}}

\item{path}{output path for files}
}
\value{
data frame of hourly fire emissions converted to units required by SMOKE. As a
  side effect, writes the ida and ems95 files to the specified path.
}
\description{
This takes fully processed hourly files, including PM2.5 emissions and plume top
estimations and produces the files needed for the SMOKE preprocessor for CMAQ. This
includes estimating emissions of other species (PM10, CO, NH3, NOX, SO2, VOC) based on
emission ratios to PM2.5, estimated from BlueSky output. The PM2.5 emissions are
assumed to be in kg, which is what is produced by \code{\link{feer_emissions}}.
}
\details{
See
\href{https://forum.cmascenter.org/t/hourly-fire-emissions-in-smoke-with-pre-computed-plume-rise/311}{this
thread} for some useful information on the formats, including strange required quirks.
PTOP and PBOT are in meters, emissions are in short tons, LAY1F cannot be zero, and
PTOP > PBOT.
}
