% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{process_fire_cubes}
\alias{process_fire_cubes}
\title{process_fire_cubes}
\usage{
process_fire_cubes(df, input_path, output_path, variables = c("Mask",
  "Area", "Power"), coords = NULL)
}
\arguments{
\item{df}{A data frame with the following fields: fire_name and sf_id (will be the name
of the output RDS file); sf_start and sf_end (begin and end dates of fire); xmin,
ymin, xmax, and ymax (coordinates of the fire's bounding box)}

\item{...}{Additional parameters passed to create_cube}
}
\value{
Returns the list of fire names processed
}
\description{
Takes a set of fire bounding boxes, expressed as a data frame with start and end dates
and lat/lon bounds and runs create cube on each, storing the results as RDS files - one
per data frame row
}
