% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_pthourly_frp}
\alias{get_pthourly_frp}
\title{Calculate hourly FRP and FRE for each satellite grid point, using imputed values for
fire mask values with no FRP reported. Also report fraction of total energy across all
points and time steps for each hour/point}
\usage{
get_pthourly_frp(df)
}
\arguments{
\item{df}{}
}
\description{
Calculate hourly FRP and FRE for each satellite grid point, using imputed values for
fire mask values with no FRP reported. Also report fraction of total energy across all
points and time steps for each hour/point
}
